+++ vulkan/DescriptorSet.cpp
@@ -29,7 +29,7 @@ bool DescriptorSet::Init(const std::shared_ptr<Device>& pDevice,
 	{
 		pDescriptorSetLayout->GetDeviceHandle(),
 	};
-	allocateInfo.descriptorSetCount = dsSetLayout.size();
+	allocateInfo.descriptorSetCount = (uint32_t)dsSetLayout.size();
 	allocateInfo.pSetLayouts = dsSetLayout.data();
 
 	CHECK_VK_ERROR(vkAllocateDescriptorSets(GetDevice()->GetDeviceHandle(), &allocateInfo, &m_descriptorSet));
@@ -58,7 +58,7 @@ void DescriptorSet::UpdateUniformBufferDynamic(uint32_t binding, const std::shar
 	VkDescriptorBufferInfo info = pBuffer->GetDescBufferInfo();
 	writeData[0].pBufferInfo = &info;
 
-	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), writeData.size(), writeData.data(), 0, nullptr);
+	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), (uint32_t)writeData.size(), writeData.data(), 0, nullptr);
 
 	m_resourceTable[binding].push_back(pBuffer);
 }
@@ -75,7 +75,7 @@ void DescriptorSet::UpdateUniformBuffer(uint32_t binding, const std::shared_ptr<
 	VkDescriptorBufferInfo info = pBuffer->GetDescBufferInfo();
 	writeData[0].pBufferInfo = &info;
 
-	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), writeData.size(), writeData.data(), 0, nullptr);
+	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), (uint32_t)writeData.size(), writeData.data(), 0, nullptr);
 
 	m_resourceTable[binding].push_back(pBuffer);
 }
@@ -95,7 +95,7 @@ void DescriptorSet::UpdateImage(uint32_t binding, const std::shared_ptr<Image>& 
 	info.sampler = pSampler->GetDeviceHandle();
 	writeData[0].pImageInfo = &info;
 
-	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), writeData.size(), writeData.data(), 0, nullptr);
+	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), (uint32_t)writeData.size(), writeData.data(), 0, nullptr);
 
 	m_resourceTable[binding].push_back(pImage);
 
@@ -118,7 +118,7 @@ void DescriptorSet::UpdateImage(uint32_t binding, const CombinedImage& image)
 	info.sampler = image.pSampler->GetDeviceHandle();
 	writeData[0].pImageInfo = &info;
 
-	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), writeData.size(), writeData.data(), 0, nullptr);
+	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), (uint32_t)writeData.size(), writeData.data(), 0, nullptr);
 
 	m_resourceTable[binding].push_back(image.pImage);
 
@@ -132,7 +132,7 @@ void DescriptorSet::UpdateImages(uint32_t binding, const std::vector<CombinedIma
 	writeData[0].sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
 	writeData[0].descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;;
 	writeData[0].dstBinding = binding;
-	writeData[0].descriptorCount = images.size();
+	writeData[0].descriptorCount = (uint32_t)images.size();
 	writeData[0].dstSet = GetDeviceHandle();
 
 	std::vector<VkDescriptorImageInfo> info;
@@ -150,7 +150,7 @@ void DescriptorSet::UpdateImages(uint32_t binding, const std::vector<CombinedIma
 	}
 	writeData[0].pImageInfo = info.data();
 
-	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), writeData.size(), writeData.data(), 0, nullptr);
+	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), (uint32_t)writeData.size(), writeData.data(), 0, nullptr);
 }
 
 void DescriptorSet::UpdateInputImage(uint32_t binding, const std::shared_ptr<Image>& pImage, const std::shared_ptr<Sampler> pSampler, const std::shared_ptr<ImageView> pImageView)
@@ -169,7 +169,7 @@ void DescriptorSet::UpdateInputImage(uint32_t binding, const std::shared_ptr<Ima
 
 	writeData[0].pImageInfo = &info;
 
-	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), writeData.size(), writeData.data(), 0, nullptr);
+	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), (uint32_t)writeData.size(), writeData.data(), 0, nullptr);
 
 	m_resourceTable[binding].push_back(pImage);
 
@@ -187,7 +187,7 @@ void DescriptorSet::UpdateTexBuffer(uint32_t binding, const VkBufferView& texBuf
 	writeData[0].dstSet = GetDeviceHandle();
 	writeData[0].pTexelBufferView = &texBufferView;
 
-	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), writeData.size(), writeData.data(), 0, nullptr);
+	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), (uint32_t)writeData.size(), writeData.data(), 0, nullptr);
 }
 
 void DescriptorSet::UpdateShaderStorageBufferDynamic(uint32_t binding, const std::shared_ptr<ShaderStorageBuffer>& pBuffer)
@@ -202,7 +202,7 @@ void DescriptorSet::UpdateShaderStorageBufferDynamic(uint32_t binding, const std
 	VkDescriptorBufferInfo info = pBuffer->GetDescBufferInfo();
 	writeData[0].pBufferInfo = &info;
 
-	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), writeData.size(), writeData.data(), 0, nullptr);
+	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), (uint32_t)writeData.size(), writeData.data(), 0, nullptr);
 
 	m_resourceTable[binding].push_back(pBuffer);
 }
@@ -219,7 +219,7 @@ void DescriptorSet::UpdateShaderStorageBuffer(uint32_t binding, const std::share
 	VkDescriptorBufferInfo info = pBuffer->GetDescBufferInfo();
 	writeData[0].pBufferInfo = &info;
 
-	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), writeData.size(), writeData.data(), 0, nullptr);
+	vkUpdateDescriptorSets(GetDevice()->GetDeviceHandle(), (uint32_t)writeData.size(), writeData.data(), 0, nullptr);
 
 	m_resourceTable[binding].push_back(pBuffer);
 }
